catalog controller.

+++ Services/Catalog/Api/IntegrationEvents/Events/ProductPriceChangedIntegrationEvent.cs
@@ -8,7 +8,7 @@ namespace ShoppDog.Services.Catalog.Api.IntegrationEvents.Events
         public decimal NewPrice { get; private set; }
         public decimal OldPrice { get; private set; }
 
-        public ProductPriceChangedIntegrationEvent(int productId, decimal newPrice, int oldPrice)
+        public ProductPriceChangedIntegrationEvent(int productId, decimal newPrice, decimal oldPrice)
         {
             ProductId = productId;
             NewPrice = newPrice;
